Extract AnnotationsOverlay component
