+++ src/lib/OpenSeadragonCanvasOverlay.js
@@ -12,23 +12,23 @@ export default class OpenSeadragonCanvasOverlay {
   /**
    * constructor - sets up the Canvas overlay container
    */
-  constructor(viewer) {
+  constructor(viewer, ref) {
     this.viewer = viewer;
+    this.ref = ref;
 
     this.containerWidth = 0;
     this.containerHeight = 0;
-
-    this.canvasDiv = document.createElement('div');
-    this.canvasDiv.style.position = 'absolute';
-    this.canvasDiv.style.left = 0;
-    this.canvasDiv.style.top = 0;
-    this.canvasDiv.style.width = '100%';
-    this.canvasDiv.style.height = '100%';
-    this.viewer.canvas.appendChild(this.canvasDiv);
-
-    this.canvas = document.createElement('canvas');
-    this.canvasDiv.appendChild(this.canvas);
     this.imgAspectRatio = 1;
+  }
+
+  /** */
+  get canvas() {
+    return this.canvasDiv.firstElementChild;
+  }
+
+  /** */
+  get canvasDiv() {
+    return this.ref.current;
   }
 
   /** */
@@ -38,7 +38,7 @@ export default class OpenSeadragonCanvasOverlay {
 
   /** */
   clear() {
-    this.canvas.getContext('2d').clearRect(0, 0, this.containerWidth, this.containerHeight);
+    this.context2d.clearRect(0, 0, this.containerWidth, this.containerHeight);
   }
 
   /**
@@ -90,11 +90,11 @@ export default class OpenSeadragonCanvasOverlay {
     ) * this.containerHeight;
 
     if (this.clearBeforeRedraw) this.clear();
-    this.canvas.getContext('2d').translate(x, y);
-    this.canvas.getContext('2d').scale(zoom, zoom);
+    this.context2d.translate(x, y);
+    this.context2d.scale(zoom, zoom);
 
     update();
 
-    this.canvas.getContext('2d').setTransform(1, 0, 0, 1, 0, 0);
+    this.context2d.setTransform(1, 0, 0, 1, 0, 0);
   }
 }
